Aggiunta del chatbot alla homepage e aggiunta di tutti i percorsi del bot
